api/requirements.txt
coverage~=4.5.3
Django~=2.2.0
django-cors-headers~=3.0.2
djangorestframework~=3.9.3
httpretty~=0.9.0
mysqlclient~=1.4.0
psl-dns~=1.0rc2
requests~=2.21.0
uwsgi~=2.0.0
